+++ lib/model/star.dart
@@ -5,9 +5,9 @@ import 'package:flutter/painting.dart';
 /// Each single [Star] which will be part the starfield.
 class Star {
   /// The base velocity the [Star] will travel at.
-  final double baseVelocity;
+  final double baseDepth;
 
-  final double spaceDepth;
+  final double baseVelocity;
 
   /// The position of the [Star] on the given x-axis.
   double dx;
@@ -22,8 +22,8 @@ class Star {
   ///
   /// It will be used to compound the starfield.
   Star({
+    required this.baseDepth,
     required this.baseVelocity,
-    required this.spaceDepth,
     required this.dx,
     required this.dy,
     required this.dz,