+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.0
 
 - Migrate to null safety
+- Improve usability of API using relative values for `depth` and `velocity`.
 
 ## 0.0.1